+++ lib.rs
@@ -51,3 +51,5 @@ impl std::fmt::Display for Error {
         }
     }
 }
+
+impl std::error::Error for Error {}